@@ -6,13 +6,15 @@ authors = ["91fabbai91 <fabian.baier@gmx.de>"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.10.12"
 httplib2 = "^0.20.4"
 lxml = "^4.9.1"
 PyYAML = "^6.0"
 pytest = "^7.1.3"
 pytest-mock = "^3.8.2"
 bidict = "^0.22.0"
+pydantic = "^1.10.7"
+python-dotenv = "^1.0.0"
 
 
 [build-system]
