@@ -1,19 +1,16 @@
-class LoginData(object):
-    def __init__(self, server: int, username: str, password: str):
-        if(server<0 or server>46):
+from pydantic import SecretStr, BaseModel,validator, Field, PositiveInt, StrictStr
+
+class LoginData(BaseModel):
+    server: PositiveInt = Field(..., allow_mutation=False)
+    username: StrictStr = Field(..., allow_mutation=False)
+    password: SecretStr = Field(..., allow_mutation=False)
+
+    class Config:
+        validate_assignment = True
+        frozen = True
+
+    @validator('server')
+    def server_match(cls, v):
+        if v < 0 or v > 46:
             raise ValueError("Server must be Integer, greater than 0 and smaller than 47")
-        self.__server = server
-        self.__username = username
-        self.__password = password
-
-    @property
-    def server(self):
-        return self.__server
-
-    @property
-    def username(self):
-        return self.__username
-
-    @property
-    def password(self):
-        return self.__password
+        return v
