@@ -1,6 +1,8 @@
 import logging
 from enum import Enum, unique
+from pydantic import BaseModel
 import http_connection
+
 
 @unique
 class BeesTour(Enum):
@@ -9,7 +11,7 @@ class BeesTour(Enum):
     TWENTYFOUR_HOURS_TOUR =3
 
 
-class BeesFarm(object):
+class BeesFarm(BaseModel):
     def __init__(self, http_connection: http_connection.HTTPConnection) -> None:
         self.__http_connection = http_connection
         self.__logger = logging.getLogger(self.__class__.__name__)
