@@ -1,23 +1,16 @@
-import http_connection
+from pydantic import BaseModel, PositiveInt, PositiveFloat
 
-class Wimp(object):
-    def __init__(self, id: int, product_amount: list, reward: float) -> None:
-        self.__http_connection = http_connection
-        self.__product_amount = product_amount
-        self.__id = id
-        self.__reward = reward
-    
-    @property
-    def product_amount(self):
-        return self.__product_amount
+class Wimp(BaseModel):
+    id: PositiveInt
+    product_amount: dict
+    reward: PositiveFloat
 
-    @property
-    def id(self):
-        return self.__id
-    
-    @property
-    def reward(self):
-        return self.__reward
+
+
+    class Config:
+        frozen = True
+        validate_assignment = True
+
 
 
     
